backend/requirements.txt
fastapi
uvicorn
openai
playwright
python-dotenv
anyio